requirements.txt
-e git://github.com/rougier/freetype-py.git#egg=freetype-py
numpy
pillow
